backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["caoguanjie <https://github.com/caoguanjie>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11.5"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
websockets = "^12.0"
openai = "^1.14.2"
python-dotenv = "^1.0.1"
pydantic = "^1.10.8"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
